restructure and stubbing out the new file utilities
